now getting hours updates and tracking what has been sent
we are now getting the list of users in the team and looping through
those and getting their updates from the workfront API, since this
method includes and hours updates as well as notes.  We are also
tracking what has already been sent to slack in a local text file.  Will
need to add a purging strategy next.